+++ plugin/output/kafka/README.md
@@ -151,7 +151,7 @@ SASL OAUTHBEARER config. It works only if `sasl_mechanism:"OAUTHBEARER"`.
 * **`client_secret`** *`string`* - client secret
 * **`token_url`** *`string`* - resource server's token endpoint URL
 * **`scopes`** *`[]string`* - optional requested permissions
-* **`auth_style`** *`string`* - specifies how the endpoint wants the client ID & client secret sent
+* **`auth_style`** *`string`* *`default=params`* *`options=params|header`* - specifies how the endpoint wants the client ID & client secret sent
 
 <br>
